+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ not released yet
 
 * FIX no more crashes when using timezones that have a constant UTC offset (like
   UTC itself)
+* in ikhal, update the date-titles on date change
 
 0.9.1
 =====